Split unit test from end-to-end tests; added commands to run unit and end-to-end tests without compiling the dependencies
