+++ package.json
@@ -5,21 +5,26 @@
     "description": "Arbitration Outsourcing for State Channels",
     "main": "index.js",
     "scripts": {
-        "build": "rm -rf ./build && npm run compile-sol && npx tsc && cp -r ./raiden_demo build/raiden_demo && cp ./src/public/favicon.ico ./build/src/public/favicon.ico",
+        "build": "rm -rf ./build && npm run build-sol && npx tsc && cp -r ./raiden_demo build/raiden_demo && cp ./src/public/favicon.ico ./build/src/public/favicon.ico",
         "build-autominer": "cd build/raiden_demo/autominer && npm install && npm run build",
         "build-client": "cd client && npm install",
         "build-raiden-pisa-daemon": "cd build/raiden_demo/raiden-pisa-daemon && npm install && npm run build",
-        "compile-sol": "cd sol && truffle compile",
+        "build-sol": "cd sol && truffle compile",
         "lint": "eslint --ext ts .",
         "start": "node ./build/src/startUp.js",
         "start-dev": "ts-node src/startUp.ts",
         "start-evm": "ganache-cli -d -l 15000000 --allowUnlimitedContractSize --defaultBalanceEther 1000000 --mnemonic 'myth like bonus scare over problem client lizard pioneer submit female collect'",
         "submodule": "git submodule init && git submodule update",
-        "test": "npm run test-deps && NODE_ENV=test mocha -r ts-node/register 'test/src/**/*.test.ts' --exit",
-        "test-alpha": "npm run test-deps && mocha -r ts-node/register test/smoke/alpha.test.ts --exit",
+        "test-deps": "npm run build-sol && npm run build-client",
+        "test": "npm run test-deps && npm run test-run-unit && npm run test-endtoend",
+        "test-run": "npm run test-run-unit && npm run test-run-endtoend",
+        "test-unit": "npm run test-deps && npm run test-run-unit",
+        "test-run-unit": "NODE_ENV=test mocha -r ts-node/register 'test/src/**/*.test.ts' --exit",
+        "test-endtoend": "NODE_ENV=test mocha -r ts-node/register test/endtoend/**/*.test.ts --exit",
+        "test-run-endtoend": "NODE_ENV=test mocha -r ts-node/register test/endtoend/**/*.test.ts --exit",
+        "test-alpha": "npm run test-deps && NODE_ENV=test mocha -r ts-node/register test/smoke/alpha.test.ts --exit",
         "test-contract": "npm run test-deps && cd sol && truffle test",
         "test-coverage": "npm run test-deps && NODE_ENV=test nyc mocha -r ts-node/register --reporter mocha-junit-reporter --reporter-options mochaFile=coverage/mocha/test-results.xml 'test/src/**/*.test.ts' --exit",
-        "test-deps": "npm run compile-sol && npm run build-client",
         "test-integration": "npm run test-deps && mocha -r ts-node/register 'test/integration/remote/**/*.test.ts' --exit",
         "test-load": "npm run test-deps && ts-node test/load/loadTest.ts 10 20000",
         "test-raiden": "npm run test-deps && mocha -r ts-node/register 'test/integration/raiden.test.ts' --exit",
